+++ Genius/Extensions/ObjectiveC/Selector.swift
@@ -1,0 +1,13 @@
+//
+// © 2026 F1248 <f1248@mailbox.org>
+// See LICENSE.txt for license information.
+//
+
+import ObjectiveC
+
+extension Selector {
+
+	init(private name: String) {
+		self.init(name)
+	}
+}